Reorganize folder structure for stories and shared components
