+++ src/components/stories/NeynarAuthButton/icons/WarpcastIcon.tsx
